CSS Modules work at runtime
