+++ build.js
@@ -24,3 +24,28 @@ rollup({
 }).then( function(bundle) {
 	bundle.write({ dest: 'dist/sofe.js', format: 'cjs' })
 });
+
+// rollup({
+// 	entry: './lib/sofe-css.js',
+// 	plugins: [
+// 		npm({
+// 			// use "jsnext:main" if possible
+// 			// – see https://github.com/rollup/rollup/wiki/jsnext:main
+// 			jsnext: true,
+//
+// 			// use "main" field or index.js, even if it's not an ES6 module
+// 			// (needs to be converted from CommonJS to ES6
+// 			// – see https://github.com/rollup/rollup-plugin-commonjs
+// 			main: true,
+//
+// 			// if there's something your bundle requires that you DON'T
+// 			// want to include, add it to 'skip'
+// 			skip: [ 'cssnano' ],
+// 		}),
+// 		commonjs()
+// 	]
+// }).then(function(bundle) {
+// 	bundle.write({ dest: 'dist/sofe-css.js', format: 'cjs' })
+// }).catch(function(error) {
+// 	console.error(error);
+// });
